Update Firebase Profile
